fastmerge

+++ X11/xc/extras/freetype2/src/pcf/pcfutil.h
@@ -23,7 +23,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
-/* $XFree86: xc/extras/freetype2/src/pcf/pcfutil.h,v 1.5 2004/04/26 16:15:54 dawes Exp $ */
+
 
 #ifndef __PCFUTIL_H__
 #define __PCFUTIL_H__
@@ -37,15 +37,15 @@ FT_BEGIN_HEADER
 
   FT_LOCAL( void )
   BitOrderInvert( unsigned char*  buf,
-                  int             nbytes );
+                  size_t          nbytes );
 
   FT_LOCAL( void )
   TwoByteSwap( unsigned char*  buf,
-               int             nbytes );
+               size_t          nbytes );
 
   FT_LOCAL( void )
   FourByteSwap( unsigned char*  buf,
-                int             nbytes );
+                size_t          nbytes );
 
 FT_END_HEADER
